update on scalability script, symlink to monodomain
git-svn-id: https://opencmiss.svn.sourceforge.net/svnroot/opencmiss/cm/trunk@1548 61e36070-c886-4827-b7c0-ed9e4dc59cc8

+++ ParallelProblemAnalysis/test_scalability.rb
@@ -6,34 +6,35 @@ summary = {}
 
 File.readlines('tests').each{|line|
   cmd = line.strip
-  next if cmd.empty?
+  next if cmd.empty? || cmd[0..0]=='#'
   cmd = cmd.split(' ')
   Dir.chdir(cmd[0]){
   `make allclean; make opt64 COMPILER=gnu` # compile opt.
   ns.each{|n|
   `rm -f time.tmp`
   fullcmd = "./bin/x86_64-linux/mpich2/gnu/#{cmd[0]}Example #{cmd[1..-1].join(' ')}"
-  `#{mpi_path}/mpiexec -n #{n} /usr/bin/time -a -o time.tmp -f "%e %U %S %M %x" #{fullcmd} &> /dev/null`
-  puts '-'*50, fullcmd, "#{n} PROCESSORS", '-'*50
+  mpicmd = "#{mpi_path}/mpiexec -n #{n} /usr/bin/time -a -o time.tmp -f \"%e %U %S %M %x\" #{fullcmd} &> /dev/null"
+  `#{mpicmd}`
+  puts '='*50, mpicmd, "#{n} PROCESSORS", '-'*20
   data = File.readlines('time.tmp').map{|l| 
     w,u,s,m,x = l.split(' ')
       puts "Walltime #{w} User #{u} System #{s} Memory #{m}  Exit code #{x}"
     [w,u+s,m,x]
   }
   summary[fullcmd] ||= []
-  summary[fullcmd] << [n] + data.transpose.map{|d| d.inject{|a,b|a.to_f+b.to_f}.to_f/n }
+  summary[fullcmd] << [n] + data.transpose.map{|d| d.inject{|a,b|a.to_f+b.to_f}.to_f/d.size }
+  summary[fullcmd][-1][-1] += 1 unless data.size==n # force error when unexpected number of processors
  }
  }
 }
- puts "------------------ SUMMARY ----------------"
-  summary.each{|c,d|
-    puts "COMMAND: #{c}", "Processors\tWalltime\tCPU Time\tMemory"
-    d.each{|data| n,w,t,m,x = data.map{|s| s.to_s.ljust 10 }; puts "#{n}\t#{w}\t#{t}\t#{m}\t#{x.to_i.zero? ? 'OK' : 'ERRROR!'}" }
-  }
- File.open('summary.out','w'){|f|
-  summary.each{|c,d|
-    f.puts "COMMAND: #{c}", "Processors\tWalltime\tCPU Time\tMemory"
-    d.each{|data| n,w,t,m,x = data.map{|s| s.to_s.ljust 10 }; f.puts "#{n}\t#{w}\t#{t}\t#{m}\t#{x.to_i.zero? ? 'OK' : 'ERRROR!'}" }
+ puts "=============================== SUMMARY =============================== "
+
+ File.open('summary.out','w'){|fo|
+  [fo,$stdout].each{|f|
+   summary.sort.each{|c,d|
+    f.puts '-'*50,"COMMAND: #{c}", "Processors\tWalltime\tCPU Time\tMemory"
+    d.each{|data| n,w,t,m,x = *data; f.puts "#{n}       \t#{'%.2f'%(w/60)}m     \t#{'%.2f'%(t/60)}m     \t#{'%.2f'%(m/1048576)}Gb/cpu  \t#{x.to_i.zero? ? 'OK' : 'ERROR!'}" } rescue puts "<ERROR PROCESSING DATA> #{d.inspect}"
+   }
   }
  }
